+++ blocklets/ai-kit/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.1.60 (2024-10-29)
+
+- chore: add access logger
+- chore: auto remove empty string from request
+
 ## 0.1.59 (2024-10-16)
 
 - fix: remove config watcher that not working in docker mode